Add files via upload
Objectives 4 5 and 6 completed. For number 5, I'm sure there is a more efficient way to do it, but my method still worked. However, I did run into an issue and am not sure how to approach it. For the Genres file, there is a column for ReleaseDate and VideoReleaseDate. VideoReleaseDate is an empty column because if you take a look at the MovieTitle, you can see that the date is actually included in its title and is not separated by a "|" like our other files. Thus, we need to figure out how to shift that date over to the ReleaseDate column, and then shift what is in the ReleaseDate column to the VideoReleaseDate column. Hopefully this makes sense, I can explain more in person.